airtable/requirements.txt
abstra==2.27.2
pandas
python-dotenv==1.0.1
requests==2.32.3